lingualens/requirements.txt
streamlit>=1.30.0
easyocr>=1.7.0
google-generativeai>=0.4.0
Pillow>=10.0.0
python-dotenv>=1.0.0
gTTS>=2.5.0
numpy>=1.24.0
